move brand logic to component
